+++ src/mjolnir/elevationbuilder.cc
@@ -67,47 +67,6 @@ std::vector<int8_t> encode_edge_elevation(const std::unique_ptr<valhalla::skadi:
   return encoded;
 }
 
-/**
- * Encode elevation for a bridge, tunnel, ferry.
- */
-std::vector<int8_t> encode_btf_elevation(const std::unique_ptr<valhalla::skadi::sample>& sample,
-                                         const std::vector<PointLL>& shape,
-                                         const uint32_t length,
-                                         uint32_t wayid) {
-  // Compute a uniform sampling interval along the edge based on its length.
-  double interval = sampling_interval(length);
-
-  // Sample at the first and last shape point
-  double h1 = sample->get(shape.front());
-  double h2 = sample->get(shape.back());
-
-  // Use linear interpolation from h1 to h2 along the length of the edge
-  uint32_t n = static_cast<uint32_t>(length / interval) + 1;
-  std::vector<double> heights(n);
-  heights.front() = h1;
-  float delta = (h2 - h1) / n;
-  for (uint32_t i = 1; i < n - 1; ++i) {
-    heights[i] = heights[i - 1] + delta;
-  }
-  heights.back() = h2;
-
-  // Encode the elevation.
-  bool error = false;
-  auto e = encode_elevation(heights, error);
-  if (error) {
-    double diff = 0;
-    for (size_t i = 1; i < heights.size(); i++) {
-      auto d = std::abs(heights[i] - heights[i - 1]);
-      diff = d < diff ? diff : d;
-      LOG_DEBUG("  " + std::to_string(heights[i]));
-    }
-    LOG_DEBUG("BTF edge elevation wayid = " + std::to_string(wayid) + " exceeds difference with " +
-              std::to_string(diff) + " meters.");
-    build_stats::get().increment(build_stats::kExceededElevationDiff);
-  }
-  return e;
-}
-
 void add_elevations_to_single_tile(GraphReader& graphreader,
                                    std::mutex& graphreader_lck,
                                    cache_t& cache,
@@ -218,12 +177,14 @@ void add_elevations_to_single_tile(GraphReader& graphreader,
       new_offsets[edge_info_offset] = ei_offset;
 
       // Encode elevation along the edge and add to EdgeInfo along with the mean elevation.
-      // Bridges, tunnels, ferries are special cases. Increment the new edge info offset.
+      // Skip for shortcuts as they are resolved into underlying edges before reading elevation.
+      // Skip for bridges, tunnels and ferries/rail-shuttles as elevation for them should be
+      // linearly interpolated between nodes and runtime does this.
+      // Increment the new edge info offset.
       std::vector<int8_t> encoded;
-      auto wayid = tilebuilder.edgeinfo(&directededge).wayid();
-      if (directededge.bridge() || directededge.tunnel() || directededge.use() == Use::kFerry) {
-        encoded = encode_btf_elevation(sample, shape, length, wayid);
-      } else {
+      if (!directededge.is_shortcut() && !directededge.bridge() && !directededge.tunnel() &&
+          directededge.use() != Use::kFerry && directededge.use() != Use::kRailFerry) {
+        auto wayid = tilebuilder.edgeinfo(&directededge).wayid();
         encoded = encode_edge_elevation(sample, shape, length, wayid);
       }
       ei_offset += tilebuilder.set_elevation(edge_info_offset, mean_elevation, encoded);