+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## UNRELEASED
 * **Removed**
+   * REMOVED: Unused elevation for shortucs and ferry/tunnel/bridge edges [#6340](https://github.com/valhalla/valhalla/pull/6340)
 * **Bug Fix**
 * **Enhancement**